Move to completion code (this code already imports completion specific types)
