+++ agent/src/bfg/BfgContextFetcher.test.ts
@@ -9,9 +9,9 @@ import * as rimraf from 'rimraf'
 import { afterAll, assert, beforeAll, describe, expect, it } from 'vitest'
 import * as vscode from 'vscode'
 
+import { BfgRetriever } from '../../../vscode/src/completions/context/retrievers/bfg/bfg-retriever'
 import { getCurrentDocContext } from '../../../vscode/src/completions/get-current-doc-context'
 import { initTreeSitterParser } from '../../../vscode/src/completions/test-helpers'
-import { BfgRetriever } from '../../../vscode/src/graph/bfg/bfg-retriever'
 import { Agent, initializeVscodeExtension } from '../agent'
 import { MessageHandler } from '../jsonrpc-alias'
 import * as vscode_shim from '../vscode-shim'
