+++ vscode/src/extension.common.ts
@@ -5,10 +5,10 @@ import { Configuration } from '@sourcegraph/cody-shared/src/configuration'
 import type { SourcegraphBrowserCompletionsClient } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/browserClient'
 import type { SourcegraphNodeCompletionsClient } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/nodeClient'
 
+import { BfgRetriever } from './completions/context/retrievers/bfg/bfg-retriever'
 import { CommandsController } from './custom-prompts/CommandsController'
 import { onActivationDevelopmentHelpers } from './dev/helpers'
 import { ExtensionApi } from './extension-api'
-import { BfgRetriever } from './graph/bfg/bfg-retriever'
 import type { FilenameContextFetcher } from './local-context/filename-context-fetcher'
 import type { LocalKeywordContextFetcher } from './local-context/local-keyword-context-fetcher'
 import type { SymfRunner } from './local-context/symf'
