+++ vscode/src/extension.node.ts
@@ -7,12 +7,12 @@ import { GitHistory } from '@sourcegraph/cody-shared/src/chat/recipes/git-log'
 import { SourcegraphNodeCompletionsClient } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/nodeClient'
 
 import { LocalIndexedKeywordSearch } from './chat/local-code-search'
+import { BfgRetriever } from './completions/context/retrievers/bfg/bfg-retriever'
 import { CommandsController } from './custom-prompts/CommandsController'
 import { ExtensionApi } from './extension-api'
 import { activate as activateCommon } from './extension.common'
 import { VSCODE_WEB_RECIPES } from './extension.web'
 import { initializeNetworkAgent, setCustomAgent } from './fetch.node'
-import { BfgRetriever } from './graph/bfg/bfg-retriever'
 import { FilenameContextFetcher } from './local-context/filename-context-fetcher'
 import { LocalKeywordContextFetcher } from './local-context/local-keyword-context-fetcher'
 import { SymfRunner } from './local-context/symf'
